add_lifecycler_event_for_default_anправки пра
