+++ mvi-core/src/main/kotlin/ru/surfstudio/mvi/lifecycle/MapperLifecycleEvent.kt
@@ -13,22 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ru.surfstudio.mvi.vm.android
+package ru.surfstudio.mvi.lifecycle
 
 import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleEventObserver
+import androidx.lifecycle.LifecycleOwner
 import ru.surfstudio.mvi.core.event.Event
 import ru.surfstudio.mvi.vm.MviViewModel
+import ru.surfstudio.mvi.vm.android.MviView
 
 /**
  * Map [Lifecycle.Event] from lifecycleScope to [LifecycleMviEvent]
  */
 interface MapperLifecycleEvent<E : Event> {
     fun mapToLifecycleScreenEvent(event: Lifecycle.Event): E
-}
-
-/**
- * Getting the function of converting a lifecycle event into a screen event
- */
-fun <E : Event> MviViewModel<E>.getMapperLifecycleEvent(): ((Lifecycle.Event) -> E)? {
-    return (this as? MapperLifecycleEvent<E>)?.let { it::mapToLifecycleScreenEvent }
 }
